requirements/prod.txt
alembic==1.4.3
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
environs==9.3.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.9.0
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Static-Digest==0.2.1
Flask-WTF==0.14.3
gevent==20.12.1
greenlet==1.0.0
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
supervisor==4.2.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.2.0
